LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/drawing-rectangles-to-the-canvas/2d.strokeRect.globalcomposite.worker-expected.txt


PASS strokeRect is not affected by globalCompositeOperation 
